WebSecurity ativado

+++ src/main/java/com/ifrn/sisgestaohospitalar/controller/AtendimentoProcedimentoController.java
@@ -1,10 +1,12 @@
 package com.ifrn.sisgestaohospitalar.controller;
 
+import java.security.Principal;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
 import javax.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.BindingResult;
@@ -16,6 +18,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import com.ifrn.sisgestaohospitalar.model.Atendimento;
 import com.ifrn.sisgestaohospitalar.model.AtendimentoProcedimento;
 import com.ifrn.sisgestaohospitalar.repository.AtendimentoRepository;
+import com.ifrn.sisgestaohospitalar.repository.ProfissionalRepository;
 import com.ifrn.sisgestaohospitalar.repository.AtendimentoProcedimentoRepository;
 
 @Controller
@@ -26,10 +29,12 @@ public class AtendimentoProcedimentoController {
 	private AtendimentoRepository atendimentoRepository;
 	@Autowired
 	private AtendimentoProcedimentoRepository atendimentoProcedimentoRepository;
+	@Autowired
+	private ProfissionalRepository profissionalRepository;
 
 	@PostMapping("/adicionar")
 	public ResponseEntity<?> adicionarProcedimentos(@Valid AtendimentoProcedimento atendimentoProcedimento,
-			BindingResult result) {
+			BindingResult result, Principal principal) {
 		Map<String, String> errors = new HashMap<>();
 		if (result.hasErrors()) {
 			for (FieldError error : result.getFieldErrors()) {
@@ -40,6 +45,7 @@ public class AtendimentoProcedimentoController {
 		Optional<Atendimento> optional = atendimentoRepository.findById(atendimentoProcedimento.getIdAtendimento());
 		if (optional.isPresent()) {
 			Atendimento atendimento = optional.get();
+			atendimentoProcedimento.setProfissional(profissionalRepository.findByCpf(principal.getName()));
 			atendimento.getAtendimentoProcedimentos().add(atendimentoProcedimento);
 			atendimentoRepository.saveAndFlush(atendimento);
 			return ResponseEntity.ok().body(atendimento);
@@ -49,7 +55,7 @@ public class AtendimentoProcedimentoController {
 
 	@GetMapping("/excluir/{idAtendimento}/{idAtendimentoProcedimento}")
 	public ResponseEntity<?> excluir(@PathVariable("idAtendimento") Long idAtendimento,
-			@PathVariable("idAtendimentoProcedimento") Long idAtendimentoProcedimento) {
+			@PathVariable("idAtendimentoProcedimento") Long idAtendimentoProcedimento, Principal principal) {
 		Optional<Atendimento> optionalAtendimento = atendimentoRepository.findById(idAtendimento);
 		if (optionalAtendimento.isPresent()) {
 			Atendimento atendimento = optionalAtendimento.get();
@@ -57,6 +63,9 @@ public class AtendimentoProcedimentoController {
 					.findById(idAtendimentoProcedimento);
 			if (optionalRelAtendimentoProcedimento.isPresent()) {
 				AtendimentoProcedimento atendimentoProcedimento = optionalRelAtendimentoProcedimento.get();
+				if (atendimentoProcedimento.getProfissional().getCpf() != principal.getName()) {
+					return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+				}
 				atendimento.getAtendimentoProcedimentos().remove(atendimentoProcedimento);
 				atendimentoRepository.saveAndFlush(atendimento);
 				atendimentoProcedimentoRepository.deleteById(idAtendimentoProcedimento);
