+++ src/main/java/com/ifrn/sisgestaohospitalar/controller/PrescricaoController.java
@@ -1,5 +1,6 @@
 package com.ifrn.sisgestaohospitalar.controller;
 
+import java.security.Principal;
 import java.time.LocalDateTime;
 import java.util.HashMap;
 import java.util.Map;
@@ -8,6 +9,7 @@ import java.util.Optional;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.BindingResult;
@@ -19,11 +21,13 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.ifrn.sisgestaohospitalar.dto.PrescricaoDTO;
+import com.ifrn.sisgestaohospitalar.enums.Status;
 import com.ifrn.sisgestaohospitalar.model.Atendimento;
 import com.ifrn.sisgestaohospitalar.model.Prescricao;
 import com.ifrn.sisgestaohospitalar.model.RegistroAdministracao;
 import com.ifrn.sisgestaohospitalar.repository.AtendimentoRepository;
 import com.ifrn.sisgestaohospitalar.repository.PrescricaoRepository;
+import com.ifrn.sisgestaohospitalar.repository.ProfissionalRepository;
 
 @Controller
 @RequestMapping("/prescricao")
@@ -33,6 +37,8 @@ public class PrescricaoController {
 	private PrescricaoRepository prescricaoRepository;
 	@Autowired
 	private AtendimentoRepository atendimentoRepository;
+	@Autowired
+	private ProfissionalRepository profissionalRepository;
 
 	@GetMapping("/{id}")
 	public ResponseEntity<?> editarPrescricao(@PathVariable("id") Long id) {
@@ -45,18 +51,22 @@ public class PrescricaoController {
 
 	@DeleteMapping("/excluir/{idAtendimento}/{idPrescricao}")
 	public ResponseEntity<?> excluirPrescricao(@PathVariable("idAtendimento") Long idAtendimento,
-			@PathVariable("idPrescricao") Long idPrescricao) {
+			@PathVariable("idPrescricao") Long idPrescricao, Principal principal) {
 		Atendimento atendimento = atendimentoRepository.getOne(idAtendimento);
 		Prescricao prescricao = prescricaoRepository.getOne(idPrescricao);
+		if (atendimento.getStatus().equals(Status.FINALIZADO)) {
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+		}
 		atendimento.getPrescricoes().remove(prescricao);
 		atendimentoRepository.saveAndFlush(atendimento);
 		prescricaoRepository.delete(prescricao);
 		return ResponseEntity.ok().build();
 
 	}
-	
+
 	@PostMapping("/registro")
-	public ResponseEntity<?> salvarRegistro(@Valid RegistroAdministracao registroAdministracao, BindingResult result) {
+	public ResponseEntity<?> salvarRegistro(@Valid RegistroAdministracao registroAdministracao, BindingResult result,
+			Principal principal) {
 		System.out.println(registroAdministracao.toString());
 		Map<String, String> errors = new HashMap<>();
 		if (result.hasErrors()) {
@@ -69,6 +79,7 @@ public class PrescricaoController {
 		Optional<Prescricao> optional = prescricaoRepository.findById(registroAdministracao.getIdPrescricao());
 		if (optional.isPresent()) {
 			Prescricao prescricao = optional.get();
+			registroAdministracao.setProfissionalResponsavel(profissionalRepository.findByCpf(principal.getName()));
 			registroAdministracao.setDataAdministracao(LocalDateTime.now());
 			prescricao.getRegistrosAdministracao().add(registroAdministracao);
 			prescricaoRepository.saveAndFlush(prescricao);
